pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docmerge"
version = "1.0.0"
description = "Organize and merge documents (PDF, DOCX, images) into structured PDFs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DocMerge Contributors"}
]
keywords = [
    "pdf",
    "document",
    "merge",
    "organize",
    "docx",
    "converter",
    "course-materials"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
dependencies = [
    "pypdf>=3.0.0",
    "pillow>=9.0.0",
    "python-docx>=0.8.11",
    "reportlab>=3.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "flake8",
]

[project.urls]
Homepage = "https://github.com/gigwebs/docmerge"
Repository = "https://github.com/gigwebs/docmerge"
Issues = "https://github.com/gigwebs/docmerge/issues"

[project.scripts]
docmerge = "docmerge.cli:main"

[tool.setuptools.packages.find]
include = ["docmerge*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
